refactor: decouple storage layer using repository pattern
