+++ server/queue.ts
@@ -33,7 +33,7 @@ export const redisConnection = new IORedis(process.env.REDIS_URL || "redis://loc
 });
 
 export const assessmentQueue = new Queue("assessmentQueue", {
-  connection: redisConnection,
+  connection: redisConnection as any,
 });
 
 const execFileAsync = promisify(execFile);
@@ -41,7 +41,7 @@ const execFileAsync = promisify(execFile);
 export const assessmentWorker = new Worker(
   "assessmentQueue",
   async (job: Job) => {
-    const { input, isPreview } = job.data;
+    const { input, isPreview, userId } = job.data;
     const tempFile = path.join(os.tmpdir(), `${randomUUID()}.json`);
 
     try {
@@ -89,7 +89,7 @@ export const assessmentWorker = new Worker(
     }
   },
   {
-    connection: redisConnection,
+    connection: redisConnection as any,
     concurrency: 4,
   }
 );
